Look up by exact username
