+++ venmo/cli.py
@@ -14,12 +14,11 @@ Refresh your venmo access token:
 """
 
 import argparse
-import json
 import urllib
 
 import requests
 
-from venmo import oauth, settings
+from venmo import oauth, settings, user
 
 
 def pay(args):
@@ -34,11 +33,15 @@ def charge(args):
 def _pay_or_charge(args):
     params = {
         'note': args.note,
-        'phone': args.phone,
         'amount': args.amount,
         'access_token': oauth.get_access_token(),
         'audience': 'private',
     }
+    if args.user.startswith("@"):
+        user_id = user.id_from_username(args.user[1:])
+        params['user_id'] = user_id
+    else:
+        params['phone'] = args.user
     response = requests.post(
         _payments_url_with_params(params)
     ).json()
@@ -86,30 +89,6 @@ def _log_response(response):
     )
 
 
-def print_search(args):
-    print json.dumps(search(args), indent=4)
-
-
-def search(args):
-    response = requests.get(
-        settings.USERS_URL,
-        params={
-            'limit': 5,
-            'query': args.query,
-        }
-    )
-    users = response.json()['data']
-    results = []
-    for user in users:
-        results.append({
-            'id': user['id'],
-            'username': user['username'],
-            'display_name': user['display_name'],
-            'profile_picture_url': user['profile_picture_url'],
-        })
-    return results
-
-
 def main():
     parser = argparse.ArgumentParser(
         description=__doc__,
@@ -119,7 +98,10 @@ def main():
 
     for action in ["pay", "charge"]:
         subparser = subparsers.add_parser(action)
-        subparser.add_argument("phone", help="who to {}".format(action))
+        subparser.add_argument(
+            "user",
+            help="who to {}, either phone or username".format(action),
+        )
         subparser.add_argument("amount", help="how much to pay or charge")
         subparser.add_argument("note", help="what the request is for")
         subparser.set_defaults(func=globals()[action])
@@ -129,7 +111,7 @@ def main():
 
     parser_search = subparsers.add_parser('search')
     parser_search.add_argument("query", help="search query")
-    parser_search.set_defaults(func=print_search)
+    parser_search.set_defaults(func=user.print_search)
 
     args = parser.parse_args()
     args.func(args)
